+++ CHANGES.rst
@@ -5,6 +5,9 @@ No longer startup MongoDBInstance with
 ``textSearchEnabled=true``, fixing startup on MongoDB 3.4
 and dropping implicit support for text search on MongoDB 2.4.
 
+#7: Oplog tool now supports MongoDB 3.4 for the tested
+use cases.
+
 5.4
 ===